+++ src/main/java/io/github/ngspace/hudder/v2runtime/values/constants/V2Array.java
@@ -1,4 +1,4 @@
-package io.github.ngspace.hudder.v2runtime.values;
+package io.github.ngspace.hudder.v2runtime.values.constants;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,12 +6,13 @@ import java.util.List;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.v2runtime.AV2Compiler;
 import io.github.ngspace.hudder.v2runtime.V2Runtime;
+import io.github.ngspace.hudder.v2runtime.values.AV2Value;
 
 public class V2Array extends AV2Value {
 
 	private AV2Value[] values;
 
-	protected V2Array(String[] strings, AV2Compiler compiler, V2Runtime runtime, int line, int charpos, String debugvalue)
+	public V2Array(String[] strings, AV2Compiler compiler, V2Runtime runtime, int line, int charpos, String debugvalue)
 			throws CompileException {
 		super(line, charpos, debugvalue, compiler);
 		values = new AV2Value[strings.length];
