+++ src/main/java/io/github/ngspace/hudder/v2runtime/values/modifiable/V2SetValue.java
@@ -1,7 +1,8 @@
-package io.github.ngspace.hudder.v2runtime.values;
+package io.github.ngspace.hudder.v2runtime.values.modifiable;
 
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.v2runtime.AV2Compiler;
+import io.github.ngspace.hudder.v2runtime.values.AV2Value;
 
 public class V2SetValue extends AV2Value {
 	public AV2Value key;
