+++ src/main/java/io/github/ngspace/hudder/v2runtime/values/operations/V2PostIncDecOperator.java
@@ -1,14 +1,15 @@
-package io.github.ngspace.hudder.v2runtime.values;
+package io.github.ngspace.hudder.v2runtime.values.operations;
 
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.v2runtime.AV2Compiler;
+import io.github.ngspace.hudder.v2runtime.values.AV2Value;
 
-public class V2PreIncDecOperator extends AV2Value {
+public class V2PostIncDecOperator extends AV2Value {
 	private AV2Value key;
 	private AV2Compiler compilerv2;
 	private boolean add;
 
-	protected V2PreIncDecOperator(AV2Value value, AV2Compiler compilerv2, int line, int charpos, boolean add,
+	public V2PostIncDecOperator(AV2Value value, AV2Compiler compilerv2, int line, int charpos, boolean add,
 			String debugvalue) {
 		super(line, charpos, debugvalue, compilerv2);
 		this.key = value;
@@ -17,12 +18,12 @@ public class V2PreIncDecOperator extends AV2Value {
 	}
 
 	@Override public Object get() throws CompileException {
-		double val = add ? key.asDouble() + 1 : key.asDouble() -1;
-		key.setValue(compilerv2, val);
+		double val = key.asDouble();
+		key.setValue(compilerv2, add ? key.asDouble() + 1 : key.asDouble() -1);
 		return val;
 	}
 	@Override public boolean isConstant() throws CompileException {return key.isConstant();}
 	@Override public void setValue(AV2Compiler compiler, Object value) throws CompileException {
-		throw new CompileException("Can't change the value of a pre-inc-dec operator", line, charpos);	
+		throw new CompileException("Can't change the value of a post-inc-dec operator", line, charpos);	
 	}
 }
