+++ src/main/java/io/github/ngspace/hudder/v2runtime/values/operations/booloperations/V2OppositeOperator.java
@@ -1,13 +1,14 @@
-package io.github.ngspace.hudder.v2runtime.values;
+package io.github.ngspace.hudder.v2runtime.values.operations.booloperations;
 
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.v2runtime.AV2Compiler;
+import io.github.ngspace.hudder.v2runtime.values.AV2Value;
 
 public class V2OppositeOperator extends AV2Value {
 	
 	private AV2Value bool;
 
-	protected V2OppositeOperator(AV2Value value, int line, int charpos, String debugvalue, AV2Compiler compiler) {
+	public V2OppositeOperator(AV2Value value, int line, int charpos, String debugvalue, AV2Compiler compiler) {
 		super(line, charpos, debugvalue, compiler);
 		this.bool = value;
 	}
